requirements.txt
protobuf==3.7.1
pysodium==0.7.1
